refactor: DTO 요청/응답 디렉터리 구분
